pyproject.toml
[project]
name = "downloads"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "webvtt-py>=0.5.1",
]
